jobs/builtin: consolidate SP_RUNNING / MAXJOBS_ALLOC / MAX_PIPESTATS / ZEXIT_NORMAL duplicates
Three more duplicate-const consolidations into zsh_h.rs re-exports:

  jobs.rs::SP_RUNNING     -> zsh_h.rs (both i32 = -1)
  jobs.rs::MAX_PIPESTATS  -> zsh_h.rs (both usize = 256)
  jobs.rs::MAXJOBS_ALLOC  -> zsh_h.rs (TYPE FIXED: was i32 in zsh_h.rs,
                                       usize in jobs.rs — both used as
                                       Vec capacity, so unified at usize)
  builtin.rs::ZEXIT_NORMAL -> zsh_h.rs (both i32 = 0)

Values matched in all four cases but the duplicate sites opened the
same drift hazard pattern as the HIST_* / TERM_UNKNOWN bit-value
bugs caught earlier in this series. Single source of truth in
zsh_h.rs keeps the entire project safe from future C-source drift.

+++ src/ported/zsh_h.rs
@@ -2179,7 +2179,12 @@ pub const STAT_DISOWN: i32 = 0x10000; // c:1095
 pub const SP_RUNNING: i32 = -1; // c:1097
 
 pub const JOBTEXTSIZE: usize = 80; // c:1104
-pub const MAXJOBS_ALLOC: i32 = 50; // c:1107
+// C: `#define MAXJOBS_ALLOC 50` (Src/zsh.h:1107) — an int literal.
+// Stored as `usize` so callers using it for Vec capacity / slice
+// indexing don't need `as usize` casts everywhere. Matches the
+// adjacent `MAX_PIPESTATS: usize` type choice (both are array
+// sizes in C).
+pub const MAXJOBS_ALLOC: usize = 50; // c:1107
 pub const MAX_PIPESTATS: usize = 256; // c:1166
 
 #[allow(non_camel_case_types)]
